+++ src/Ydb.Sdk/src/Pool/ChannelPool.cs
@@ -103,7 +103,8 @@ internal class GrpcChannelFactory : IChannelFactory<GrpcChannel>
             // https://github.com/grpc/proposal/blob/master/A8-client-side-keepalive.md
             KeepAlivePingDelay = _config.KeepAlivePingDelay,
             KeepAlivePingTimeout = _config.KeepAlivePingTimeout,
-            KeepAlivePingPolicy = HttpKeepAlivePingPolicy.Always
+            KeepAlivePingPolicy = HttpKeepAlivePingPolicy.Always,
+            EnableMultipleHttp2Connections = _config.EnableMultipleHttp2Connections
         };
 
         // https://github.com/grpc/grpc-dotnet/issues/2312#issuecomment-1790661801